FIX:	ex00 should be done left the format of floats and doubles as is

+++ cpp06/ex00/ScalarConverter.hpp
@@ -6,7 +6,7 @@
 /*   By: pharbst <pharbst@student.42heilbronn.de>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/14 11:46:05 by pharbst           #+#    #+#             */
-/*   Updated: 2023/09/19 17:12:26 by pharbst          ###   ########.fr       */
+/*   Updated: 2023/09/20 14:44:31 by pharbst          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,16 +17,15 @@
 # include <cmath>
 # include <iostream>
 # include <string>
+# include <climits>
 # include <limits>
-# include <iomanip>
-# include <sstream>
 
 class ScalarConverter {
 	private:
 		ScalarConverter();
 		~ScalarConverter();
 		static void			_char(char c);
-		static void			_int(int i);
+		static void			_int(long i);
 		static void			_float(float f);
 		static void			_double(double d);
 	public:
